+++ wordcount/wordcount/FileReadAllLinesBenchmark.cs
@@ -6,7 +6,7 @@ public static class FileReadAllLinesBenchmark
 {
     public static Count Count(string path)
     {
-        int wordCount = 0, lineCount = 0, charCount = 0;
+        long wordCount = 0, lineCount = 0, charCount = 0;
 
         foreach (var line in File.ReadAllLines(path))
         {
